+++ tests/testthat/test-esp-get-landwater.R
@@ -2,7 +2,9 @@ test_that("Test offline", {
   skip_on_cran()
   skip_if_siane_offline()
 
-  options(mapspain_test_offline = TRUE)
+  local_mocked_bindings(is_online_fun = function(...) {
+    FALSE
+  })
   expect_message(
     n <- esp_get_rivers(update_cache = TRUE),
     "Offline"
@@ -19,14 +21,18 @@ test_that("Test offline", {
   )
   expect_null(n)
 
-  options(mapspain_test_offline = FALSE)
+  local_mocked_bindings(is_online_fun = function(...) {
+    httr2::is_online()
+  })
 })
 
 test_that("Test 404", {
   skip_on_cran()
   skip_if_siane_offline()
 
-  options(mapspain_test_404 = TRUE)
+  local_mocked_bindings(is_404 = function(...) {
+    TRUE
+  })
   expect_message(
     n <- esp_get_rivers(update_cache = TRUE),
     "Error"
@@ -42,7 +48,9 @@ test_that("Test 404", {
     "Error"
   )
   expect_null(n)
-  options(mapspain_test_404 = FALSE)
+  local_mocked_bindings(is_404 = function(...) {
+    FALSE
+  })
 })