+++ tests/testthat/test-esp-get-tiles.R
@@ -3,8 +3,9 @@ test_that("Test offline", {
   skip_if_siane_offline()
   skip_if_gisco_offline()
 
-  options(gisco_test_offline = TRUE)
-  options(mapspain_test_offline = TRUE)
+  local_mocked_bindings(is_online_fun = function(...) {
+    FALSE
+  })
   x <- esp_nuts_2024[1, ]
 
   # WMTS
@@ -31,8 +32,9 @@ test_that("Test offline", {
   )
   expect_null(n)
 
-  options(mapspain_test_offline = FALSE)
-  options(gisco_test_offline = FALSE)
+  local_mocked_bindings(is_online_fun = function(...) {
+    httr2::is_online()
+  })
 })
 
 test_that("Test 404", {
@@ -40,8 +42,9 @@ test_that("Test 404", {
   skip_if_siane_offline()
   skip_if_gisco_offline()
 
-  options(gisco_test_404 = TRUE)
-  options(mapspain_test_404 = TRUE)
+  local_mocked_bindings(is_404 = function(...) {
+    TRUE
+  })
   x <- esp_nuts_2024[1, ]
   expect_message(
     n <- esp_get_tiles(
@@ -65,8 +68,9 @@ test_that("Test 404", {
   )
   expect_null(n)
 
-  options(mapspain_test_404 = FALSE)
-  options(gisco_test_404 = FALSE)
+  local_mocked_bindings(is_404 = function(...) {
+    FALSE
+  })
 })
 
 test_that("tiles error", {