+++ NEWS.md
@@ -1,5 +1,7 @@
 # mapSpain (development version)
 
+-   Use `testthat::local_mocked_bindings()` on API error testing.
+
 # mapSpain 1.0.0
 
 This major release introduces a full overhaul of the codebase and test suite.